Added adapter. Implementing recycler in MainActivity WIP

+++ app/src/main/java/com/example/challenge3fragmentsbootcamp/Crypto.kt
@@ -3,4 +3,4 @@ package com.example.challenge3fragmentsbootcamp
 import com.example.challenge3fragmentsbootcamp.api.Image
 
 data class Crypto(val nombre: String, val abreviatura: String, val imagen: Image,
-                  val fecha_creacion: String, val precio: Long)
+                  val fecha_creacion: String, val precio: Double)
